# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie App API",
        "description": "Tài liệu API cho hệ thống đặt vé xem phim",
        "contact": {
            "email": "youremail@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/phong/dsghe/{ma_phong}": {
            "get": {
                "tags": [
                    "Phòng Chiếu"
                ],
                "summary": "Lấy danh sách ghế theo phòng chiếu",
                "description": "API này dùng để lấy danh sách các ghế trong một phòng chiếu cụ thể.",
                "operationId": "a69c4d9ad463466bcaede452fb5df872",
                "parameters": [
                    {
                        "name": "ma_phong",
                        "in": "path",
                        "description": "Mã phòng chiếu cần lấy ghế",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách ghế trong phòng chiếu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "ma_ghe": {
                                                "type": "string",
                                                "example": "G01"
                                            },
                                            "so_ghe": {
                                                "type": "string",
                                                "example": "A1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy ghế nào trong phòng chiếu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Không tìm thấy ghế nào"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/phong/{ma_phong}": {
            "get": {
                "tags": [
                    "Phòng Chiếu"
                ],
                "summary": "Lấy thông tin phòng chiếu theo mã",
                "description": "API này dùng để lấy thông tin chi tiết một phòng chiếu theo mã phòng.",
                "operationId": "f64f7fe4de224517e654e6822e0c6cb2",
                "parameters": [
                    {
                        "name": "ma_phong",
                        "in": "path",
                        "description": "Mã phòng chiếu cần tìm",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin phòng chiếu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ma_phong": {
                                            "type": "string",
                                            "example": "PC001"
                                        },
                                        "ten_phong": {
                                            "type": "string",
                                            "example": "Phòng Chiếu 1"
                                        },
                                        "so_ghe": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "mo_ta": {
                                            "type": "string",
                                            "example": "Phòng chiếu hiện đại với ghế ngồi thoải mái"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Phòng chiếu không tồn tại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phòng chiếu không tồn tại"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rap/{ma_rap}": {
            "get": {
                "tags": [
                    "Rạp chiếu phim"
                ],
                "summary": "Lấy thông tin rạp chiếu phim",
                "description": "API này dùng để lấy thông tin chi tiết của một rạp chiếu phim theo ID.",
                "operationId": "288dc31faf3ca073cc1b79008f976225",
                "parameters": [
                    {
                        "name": "ma_rap",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "RP001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thông tin rạp chiếu phim thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "RP001"
                                        },
                                        "ten": {
                                            "type": "string",
                                            "example": "Rạp Phim ABC"
                                        },
                                        "dia_chi": {
                                            "type": "string",
                                            "example": "123 Đường ABC, Quận 1, TP.HCM"
                                        },
                                        "so_phong_chieu": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rạp chiếu phim không tồn tại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Không tìm thấy rạp chiếu phim!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}